+++ R/phylospatial-package.R
@@ -4,6 +4,7 @@
 ## usethis namespace: start
 #' @importFrom stats na.omit
 #' @importFrom stats sd
+#' @importFrom stats var
 #' @importFrom stats weighted.mean
 #' @importFrom utils getFromNamespace
 ## usethis namespace: end