+++ NEWS.md
@@ -1,3 +1,5 @@
+# phylospatial (development version)
+
 # phylospatial 1.2.0
 
 * CRAN compliance: fixed vignette builds to conditionally load suggested package 'tmap'.