+++ NEWS.md
@@ -1,8 +1,8 @@
-# phylospatial (development version)
+# phylospatial 1.2.1
 
 * `ps_diversity()`, `ps_rand()`, `ps_dissim()`, and `ps_prioritize()` have been refactored to optimize compute speed (~2x to 20x speedup).
 
-* `ps_ordinate` now defaults to `method = "cmds"`, and has a bug fixed in its `pca` method.
+* `ps_ordinate()` now defaults to `method = "cmds"`, and has a bug fixed in its `"pca"` method.
 
 # phylospatial 1.2.0